requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
flask==2.3.3
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.58.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.1.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
repoze.lru==0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.21
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tflite==2.2.0
typing-extensions==4.8.0
urllib3==1.26.16
werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
